+++ src/main/java/com/github/ko4evneg/tb/bot/TelegramBot.java
@@ -1,5 +1,9 @@
 package com.github.ko4evneg.tb.bot;
 
+import com.github.ko4evneg.tb.comand.CommandContainer;
+import com.github.ko4evneg.tb.service.SendBotMessageService;
+import com.github.ko4evneg.tb.service.SendBotMessageServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
@@ -7,13 +11,26 @@ import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
+import java.util.Locale;
+
+import static com.github.ko4evneg.tb.comand.CommandName.NO;
+
 @Component
 public class TelegramBot extends TelegramLongPollingBot {
+    private final static String COMMAND_PREFIX = "/";
+
     @Value("${bot.username}")
     private String botUsername;
 
     @Value("${bot.token}")
     private String botToken;
+
+
+    private final CommandContainer commandContainer;
+
+    public TelegramBot() {
+        commandContainer = new CommandContainer(new SendBotMessageServiceImpl(this));
+    }
 
     @Override
     public String getBotToken() {
@@ -26,15 +43,11 @@ public class TelegramBot extends TelegramLongPollingBot {
             String message = update.getMessage().getText().trim();
             String chatId = update.getMessage().getChatId().toString();
 
-            SendMessage sendMessage = new SendMessage();
-            sendMessage.setChatId(chatId);
-            sendMessage.setText(message);
-
-            try {
-                execute(sendMessage);
-            } catch (TelegramApiException e) {
-                //TODO: add logging
-                e.printStackTrace();
+            if (message.startsWith(COMMAND_PREFIX)) {
+                String commandId = message.toLowerCase();
+                commandContainer.retrieveCommand(commandId).execute(update);
+            } else {
+                commandContainer.retrieveCommand(NO.getCommandName()).execute(update);
             }
         }
     }