+++ src/test/java/com/github/ko4evneg/tb/comand/AbstractCommandTest.java
@@ -1,0 +1,43 @@
+package com.github.ko4evneg.tb.comand;
+
+import com.github.ko4evneg.tb.bot.TelegramBot;
+import com.github.ko4evneg.tb.service.SendBotMessageService;
+import com.github.ko4evneg.tb.service.SendBotMessageServiceImpl;
+import org.checkerframework.checker.units.qual.C;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Message;
+import org.telegram.telegrambots.meta.api.objects.Update;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+public abstract class AbstractCommandTest {
+    protected TelegramBot telegramBot = Mockito.mock(TelegramBot.class);
+    protected SendBotMessageService sendBotMessageService = new SendBotMessageServiceImpl(telegramBot);
+
+    abstract String getCommandName();
+    abstract Command getCommand();
+    abstract String getCommandMessage();
+
+    @Test
+    public void shouldProperlyExecuteCommand() throws TelegramApiException {
+        Long chatId = 1234567824356L;
+
+        //Mocked update with message
+        Update update = new Update();
+        Message message = Mockito.mock(Message.class);
+        Mockito.when(message.getChatId()).thenReturn(chatId);
+        Mockito.when(message.getText()).thenReturn(getCommandName());
+        update.setMessage(message);
+
+        //Reference message for comparison
+        SendMessage sendMessage = new SendMessage();
+        sendMessage.setChatId(chatId.toString());
+        sendMessage.setText(getCommandMessage());
+        sendMessage.enableHtml(true);
+
+        getCommand().execute(update);
+
+        Mockito.verify(telegramBot).execute(sendMessage);
+    }
+}