+++ src/test/java/com/github/ko4evneg/tb/service/SendBotMessageServiceTest.java
@@ -1,0 +1,36 @@
+package com.github.ko4evneg.tb.service;
+
+import com.github.ko4evneg.tb.bot.TelegramBot;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+
+class SendBotMessageServiceTest {
+    private SendBotMessageService sendBotMessageService;
+    private TelegramBot telegramBot;
+
+    @BeforeEach
+    public void init() {
+        telegramBot = Mockito.mock(TelegramBot.class);
+        sendBotMessageService = new SendBotMessageServiceImpl(telegramBot);
+    }
+
+    @Test
+    public void shouldProperlySendMessage() throws TelegramApiException {
+        String chatId = "test_chat_id";
+        String message = "test_message";
+        SendMessage sendMessage = new SendMessage();
+        sendMessage.setText(message);
+        sendMessage.setChatId(chatId);
+        sendMessage.enableHtml(true);
+
+        sendBotMessageService.sendMessage(chatId, message);
+
+        Mockito.verify(telegramBot).execute(sendMessage);
+    }
+}