+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 #  Release notes
 
+###  _0.2.0-SNAPSHOT_
+- Added command patern for handling Telegram bot commands
+- Added sending service
+
 ###  _0.1.0-SNAPSHOT_
 - Added Spring Boot skeleton prokect
 - Added telegram bot